1.修复添加加盟标签存错userid的错误、2.修复获取加盟追踪标签未区分id的错误、3.修改添加加盟订单中自动添加填表人和接收前端数组、4.添加首页左下角数据统计、5.更新数据库sql文件

+++ nzj/src/org/dao/imp/JoinOrderDaoImp.java
@@ -21,7 +21,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 public class JoinOrderDaoImp implements JoinOrderDao {
 
 	@Override
-	public boolean addJoninOrder(JoinOrders jo, final long[] id) {
+	public boolean addJoninOrder(JoinOrders jo, final Long[] id) {
 		try {
 			Session session = HibernateSessionFactory.getSession();
 			Transaction ts = session.beginTransaction();
@@ -35,7 +35,7 @@ public class JoinOrderDaoImp implements JoinOrderDao {
 					conn.setAutoCommit(false);
 					for (int i = 0; i < id.length; i++) {
 						stmt.setLong(1, joId);
-						stmt.setLong(2, id[i]);
+						stmt.setLong(2, id[i].longValue());
 						stmt.addBatch();
 					}
 					stmt.executeBatch();
